Changing the program into modular, optimizing and cleaning the code part1
This will be a big change, the reason is that when I or other person wants to change something in the code of the program or add a feature to it, you have to edit some internal files of the program and if you've an error with a file the error can be related to another file its handling, this update is mainly to work with the program more easily and clean everything, so, with this update the goals are:

- Make a solid KB-E Base with "libraries" that modules can use to get determinated data

- Make easier to add a feature to the program making a single module which KB-E can read automatically, identify and use it

- Make a single file for the paths and variables for each internal file inside the program Base, this will make the modules to handle more easily their data requeriments

- Clean the whole program base code, fix possible bugs and current non-discovered bugs

Also, buildkernel.sh can be considered has a module but it's the essence of the program, so, I'll keep it inside the base. Basically, other things non-related to the building of kernel is considered a Feature (Module).